order/requirements.txt
quart==0.20.1
uvicorn==0.34.0
redis[hiredis]==5.0.3
msgspec==0.18.6
httpx==0.27.0
grpcio==1.78.0
protobuf>=6.31.1
